+++ src/backend/core/tests/documents/test_api_documents_list.py
@@ -39,9 +39,11 @@ def test_api_documents_list_format():
     client = APIClient()
     client.force_login(user)
 
+    other_users = factories.UserFactory.create_batch(3)
     document = factories.DocumentFactory(
         users=[user, *factories.UserFactory.create_batch(2)],
-        favorited_by=[user, *factories.UserFactory.create_batch(2)],
+        favorited_by=[user, *other_users],
+        link_traces=other_users,
     )
 
     response = client.get("/api/v1.0/documents/")
@@ -63,6 +65,7 @@ def test_api_documents_list_format():
         "is_favorite": True,
         "link_reach": document.link_reach,
         "link_role": document.link_role,
+        "nb_accesses": 3,
         "title": document.title,
         "updated_at": document.updated_at.isoformat().replace("+00:00", "Z"),
     }
@@ -92,9 +95,7 @@ def test_api_documents_list_authenticated_direct(django_assert_num_queries):
     expected_ids = {str(document.id) for document in documents}
 
     with django_assert_num_queries(3):
-        response = client.get(
-            "/api/v1.0/documents/",
-        )
+        response = client.get("/api/v1.0/documents/")
 
     assert response.status_code == 200
     results = response.json()["results"]